+++ api/mon/backends/zabbix/exceptions.py
@@ -1,0 +1,2 @@
+# noinspection PyUnresolvedReferences
+from api.mon.exceptions import *  # noqa: F401,F403