+++ api/mon/constants.py
@@ -1,0 +1,5 @@
+# Return values for methods that manipulate monitoring objects
+MON_OBJ_NOTHING = 0
+MON_OBJ_CREATED = 1
+MON_OBJ_UPDATED = 2
+MON_OBJ_DELETED = 3