+++ api/mon/vm/tasks.py
@@ -7,7 +7,8 @@ from que.exceptions import MgmtTaskException
 from que.utils import dc_id_from_task_id
 from api.task.utils import mgmt_task, mgmt_lock
 from api.mon.log import save_task_log
-from api.mon import LOG, get_monitoring, MonitoringError
+from api.mon import LOG, get_monitoring
+from api.mon.exceptions import MonitoringError
 from api.mon.messages import LOG_MON_VM_UPDATE, LOG_MON_VM_DELETE
 from api.mon.vm.utils import VmMonInternalTask
 from vms.signals import vm_deployed, vm_json_active_changed, vm_node_changed, vm_notcreated
@@ -80,12 +81,6 @@ def mon_vm_sync(task_id, sender, vm_uuid=None, log=LOG, **kwargs):
 
     if vm.is_deploying():
         logger.warn('Ignoring VM %s zabbix sync, because it is currently being deployed', vm)
-        return None
-
-    # This can be also called via vm_zoneid_changed signal, where VM's DC is not available and
-    # monitoring can be disabled in that DC
-    if not vm.dc.settings.MON_ZABBIX_ENABLED:
-        logger.info('Skipping VM %s zabbix sync, because zabbix module is completely disabled in DC %s', vm, vm.dc)
         return None
 
     zx = get_monitoring(vm.dc)