+++ gui/context_processors.py
@@ -38,9 +38,12 @@ def common_stuff(request):
     """
     Make common settings and variables available in templates.
     """
+    from gui.node.utils import get_dc1_settings
+
     return {
         'settings': settings,
         'dc_settings': request.dc.settings,
+        'dc1_settings': get_dc1_settings(request),
         'ANALYTICS': None if request.user.is_authenticated() else settings.ANALYTICS,
         'DEBUG': _get_debug_settings(),
         'SOCKETIO_URL': settings.SOCKETIO_URL,